Basic Express Implementation
Implemented Express. The Firebase Hosting App redirecting the user to other Firebase Functions associated, using express to display a dynamic website.

A basic GET request for the path "api/data/:location/opens" displaying the number of page opens stored in a Firestore Database.

The number does not update yet as only the GET handling is implemented for now.