Session Cookies Implemented
To make sure that data is logged only once for individual devices, I made a client side static web page. It will automatically sign the user in anonymously using FirebaseAuth, with no need for any sort of credentials, because FirebaseAuth will grand a session cookie for the user. Once the sign in is successful, the user is redirected to the server side backend, where a Firebase Function uses the location and UID sent in the URL parameters to check if that user has used the link from that location before. If not, it will add it in the database and update it, if not, it will inform the user that they opened the link before.